move item inteface to model
