+++ src/components/information/information-presenter.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react'
 import { Trip } from '../../tripSource'
 import InformationView from './information-view'
-import { ItemGroup } from '../profile-page/profile-page-components/profile-page-item'
+import { ItemGroup } from '../../Model'
 
 interface InformationPresenterProps {
   originAddress: string | undefined
